[3.5] Add Orchestra\Testbench\Dusk\Concerns\Dusk trait.
This class would contain minimal requirement to use chrome web driver on any local or remote browser testing.

I want a quick way to run complete application availability for my app (useful after big deployment, database maintenance). Loading and serving Laravel isn't useful here since the request is mainly remote.
